:recycle: [Refactor] 컴포넌트 위치 수정 & 폴더/확장자명 변경
