+++ src/components/RatingInfo.stories.tsx
